moved duplicate type a constants, commands and configs to a new, more generalised type_a module, not completly finished yet
