+++ src/epd1in54/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) const WIDTH: u16 = 128;
+pub(crate) const HEIGHT: u16 = 296;
+pub(crate) const DPI: u16 = 184;
